skapat showcourses component
